Init commit - torch profiling
Have a benchmarking class but it's not very useful. Haven't tried the quantized version of the model yet but the regular version is INCREDIBLY SLOW, to the point where benchmarking is quite out of the question unless I wanted to wait 2-3 hours for 30 questions to process. Probably means a 6.7B model is too large for local running, but after using the torch profiler it does appear that the GPU isn't getting the use it should, so more exploring may be needed on whether I'm just missing something.